chore(public): restructure files in the public folder
