+++ public/index.html
@@ -55,6 +55,8 @@
         </table>
     </div>
     <script type="text/javascript" src="/js/index.js"></script>
+    <script type="text/javascript" src="/js/utils/consts.js"></script>
+    <script type="text/javascript" src="/js/api/index.js"></script>
 </body>
 
 </html>
