+++ public/js/utils/consts.js
@@ -1,0 +1,1 @@
+const BASE_URL = 'http://127.0.0.1:3000/api/v1/businessDates';
