+++ public/js/utils/index.js
@@ -1,0 +1,2 @@
+import * as consts from './consts'
+
